+++ selfdrive/car/honda/interface.py
@@ -1,17 +1,19 @@
 #!/usr/bin/env python
+import os
 import time
 import common.numpy_fast as np
 
 from selfdrive.config import Conversions as CV
-from selfdrive.car.honda.carstate import CarState
-from selfdrive.car.honda.carcontroller import CarController, AH
+from .carstate import CarState
+from .carcontroller import CarController, AH
 from selfdrive.boardd.boardd import can_capnp_to_can_list
 
 from cereal import car
 
-import zmq
 from selfdrive.services import service_list
 import selfdrive.messaging as messaging
+
+NEW_CAN = os.getenv("OLD_CAN") is None
 
 # Car button codes
 class CruiseButtons:
@@ -35,7 +37,6 @@ class BP:
   TRIPLE = 2
   REPEATED = 1
 
-
 class CarInterface(object):
   def __init__(self, CP, logcan, sendcan=None):
     self.logcan = logcan
@@ -55,19 +56,96 @@ class CarInterface(object):
     if self.CS.accord:
       self.accord_msg = []
 
+  @staticmethod
+  def get_params(candidate, fingerprint):
+
+    # kg of standard extra cargo to count for drive, gas, etc...
+    std_cargo = 136
+
+    ret = car.CarParams.new_message()
+
+    ret.carName = "honda"
+    ret.radarName = "nidec"
+    ret.carFingerprint = candidate
+
+    ret.enableSteer = True
+    ret.enableBrake = True
+
+    # pedal
+    ret.enableGas = 0x201 in fingerprint
+    ret.enableCruise = not ret.enableGas
+
+    # FIXME: hardcoding honda civic 2016 touring wight so it can be used to 
+    # scale unknown params for other cars
+    m_civic = 2923./2.205 + std_cargo
+    l_civic = 2.70
+    aF_civic = l_civic * 0.4
+    aR_civic = l_civic - aF_civic
+    j_civic = 2500
+    cF_civic = 85400
+    cR_civic = 90000
+
+    if candidate == "HONDA CIVIC 2016 TOURING":
+      ret.m = m_civic
+      ret.l = l_civic
+      ret.aF = aF_civic
+      ret.sR = 13.0
+      ret.steerKp, ret.steerKi = 1.0, 0.24
+    elif candidate == "ACURA ILX 2016 ACURAWATCH PLUS":
+      ret.m = 3095./2.205 + std_cargo
+      ret.l = 2.67
+      ret.aF = ret.l * 0.37
+      ret.sR = 15.3
+      # Acura at comma has modified steering FW, so different tuning for the Neo in that car
+      # FIXME: using dongleId isn't great, better to identify the car than the Neo
+      is_fw_modified = os.getenv("DONGLE_ID") == 'cb38263377b873ee'
+      ret.steerKp, ret.steerKi = [0.5, 0.12] if is_fw_modified else [1.0, 0.24]
+    elif candidate == "HONDA ACCORD 2016 TOURING":
+      ret.m = 3580./2.205 + std_cargo
+      ret.l = 2.74
+      ret.aF = ret.l * 0.38
+      ret.sR = 15.3
+      ret.steerKp, ret.steerKi = 1.0, 0.24
+    elif candidate == "HONDA CR-V 2016 TOURING":
+      ret.m = 3572./2.205 + std_cargo
+      ret.l = 2.62
+      ret.aF = ret.l * 0.41
+      ret.sR = 15.3
+      ret.steerKp, ret.steerKi = 0.5, 0.12
+    else:
+      raise ValueError("unsupported car %s" % candidate)
+ 
+    ret.aR = ret.l - ret.aF
+    # TODO: get actual value, for now starting with reasonable value for 
+    # civic and scaling by mass and wheelbase
+    ret.j = j_civic * ret.m * ret.l**2 / (m_civic * l_civic**2)
+
+    # TODO: start from empirically derived lateral slip stiffness for the civic and scale by
+    # mass and CG position... all cars will have approximately similar dyn behaviors
+    ret.cF = cF_civic * ret.m / m_civic * (ret.aR / ret.l) / (aR_civic / l_civic)
+    ret.cR = cR_civic * ret.m / m_civic * (ret.aF / ret.l) / (aF_civic / l_civic)
+
+    # no rear steering, at least on the listed cars above
+    ret.chi = 0.
+
+    return ret
+
   # returns a car.CarState
   def update(self):
     # ******************* do can recv *******************
     can_pub_main = []
     canMonoTimes = []
 
-    for a in messaging.drain_sock(self.logcan):
-      canMonoTimes.append(a.logMonoTime)
-      can_pub_main.extend(can_capnp_to_can_list(a.can, [0,2]))
-      if self.CS.accord:
-        self.accord_msg.extend(can_capnp_to_can_list(a.can, [9]))
-        self.accord_msg = self.accord_msg[-1:]
-    self.CS.update(can_pub_main)
+    if NEW_CAN:
+      self.CS.update(can_pub_main)
+    else:
+      for a in messaging.drain_sock(self.logcan):
+        canMonoTimes.append(a.logMonoTime)
+        can_pub_main.extend(can_capnp_to_can_list(a.can, [0,0x80]))
+        if self.CS.accord:
+          self.accord_msg.extend(can_capnp_to_can_list(a.can, [9]))
+          self.accord_msg = self.accord_msg[-1:]
+      self.CS.update(can_pub_main)
 
     # create message
     ret = car.CarState.new_message()
@@ -115,6 +193,7 @@ class CarInterface(object):
     # cruise state
     ret.cruiseState.enabled = self.CS.pcm_acc_status != 0
     ret.cruiseState.speed = self.CS.v_cruise_pcm * CV.KPH_TO_MS
+    ret.cruiseState.available = bool(self.CS.main_on)
 
     # TODO: button presses
     buttonEvents = []
@@ -178,7 +257,7 @@ class CarInterface(object):
     if self.CS.steer_error:
       errors.append('steerUnavailable')
     elif self.CS.steer_not_allowed:
-      errors.append('steerTemporarilyUnavailable')
+      errors.append('steerTempUnavailable')
     if self.CS.brake_error:
       errors.append('brakeUnavailable')
     if not self.CS.gear_shifter_valid:
@@ -230,7 +309,7 @@ class CarInterface(object):
       "chimeRepeated": (BP.MUTE, CM.REPEATED),
       "chimeContinuous": (BP.MUTE, CM.CONTINUOUS)}[str(c.hudControl.audibleAlert)]
 
-    pcm_accel = int(np.clip(c.cruiseControl.accelOverride/1.4,0,1)*0xc6)
+    pcm_accel = int(np.clip(c.cruiseControl.accelOverride,0,1)*0xc6)
 
     self.CC.update(self.sendcan, c.enabled, self.CS, self.frame, \
       c.gas, c.brake, c.steeringTorque, \