+++ selfdrive/controls/lib/adaptivecruise.py
@@ -3,6 +3,8 @@ import numpy as np
 from common.numpy_fast import clip, interp
 import selfdrive.messaging as messaging
 
+# TODO: we compute a_pcm but we don't use it, as accelOverride is hardcoded in controlsd
+
 # lookup tables VS speed to determine min and max accels in cruise
 _A_CRUISE_MIN_V  = [-1.0, -.8, -.67, -.5, -.30]
 _A_CRUISE_MIN_BP = [   0., 5.,  10., 20.,  40.]
@@ -25,7 +27,7 @@ def limit_accel_in_turns(v_ego, angle_steers, a_target, a_pcm, CP):
   deg_to_rad = np.pi / 180.  # from can reading to rad
 
   a_total_max = interp(v_ego, _A_TOTAL_MAX_BP, _A_TOTAL_MAX_V)
-  a_y = v_ego**2 * angle_steers * deg_to_rad / (CP.steerRatio * CP.wheelBase)
+  a_y = v_ego**2 * angle_steers * deg_to_rad / (CP.sR * CP.l)
   a_x_allowed = math.sqrt(max(a_total_max**2 - a_y**2, 0.))
 
   a_target[1] = min(a_target[1], a_x_allowed)
@@ -109,12 +111,15 @@ _A_CORR_BY_SPEED_V = [0.4, 0.4, 0]
 # speeds
 _A_CORR_BY_SPEED_BP = [0., 5., 20.]
 
+# max acceleration allowed in acc, which happens in restart
+A_ACC_MAX = max(_A_CORR_BY_SPEED_V) + max(_A_CRUISE_MAX_V)
+
 def calc_positive_accel_limit(d_lead, d_des, v_ego, v_rel, v_ref, v_rel_ref, v_coast, v_target, a_lead_contr, a_max):
   a_coast_min = -1.0   # never coast faster then -1m/s^2
   # coasting behavior above v_coast. Forcing a_max to be negative will force the pid_speed to decrease,
   # regardless v_target
   if v_ref > min(v_coast, v_target):
-    # for smooth coast we can be agrressive and target a point where car would actually crash
+    # for smooth coast we can be aggressive and target a point where car would actually crash
     v_offset_coast = 0.
     d_offset_coast = d_des/2. - 4.
 
@@ -127,9 +132,8 @@ def calc_positive_accel_limit(d_lead, d_des, v_ego, v_rel, v_ref, v_rel_ref, v_c
     else:
       a_max = a_coast_min
   else:
-    # same as cruise accel, but add a small correction based on lead acceleration at low speeds
-    # when lead car accelerates faster, we can do the same, and vice versa
-
+    # same as cruise accel, plus add a small correction based on relative lead speed
+    # if the lead car is faster, we can accelerate more, if the car is slower, then we can reduce acceleration
     a_max = a_max + interp(v_ego, _A_CORR_BY_SPEED_BP, _A_CORR_BY_SPEED_V) \
                   * clip(-v_rel / 4., -.5, 1)
   return a_max
@@ -154,7 +158,7 @@ def calc_acc_accel_limits(d_lead, d_des, v_ego, v_pid, v_lead, v_rel, a_lead,
                           v_target, v_coast, a_target, a_pcm):
   #*** compute max accel ***
   # v_rel is now your velocity in lead car frame
-  v_rel = -v_rel  # this simplifiess things when thinking in d_rel-v_rel diagram
+  v_rel *= -1  # this simplifies things when thinking in d_rel-v_rel diagram
 
   v_rel_pid = v_pid - v_lead
 
@@ -214,28 +218,6 @@ def calc_jerk_factor(d_lead, v_rel):
   return jerk_factor
 
 
-def calc_ttc(d_rel, v_rel, a_rel, v_lead):
-  # this function returns the time to collision (ttc), assuming that a_rel will stay constant
-  # TODO: Review these assumptions.
-  # change sign to rel quantities as it's going to be easier for calculations
-  v_rel = -v_rel
-  a_rel = -a_rel
-
-  # assuming that closing gap a_rel comes from lead vehicle decel, then limit a_rel so that v_lead will get to zero in no sooner than t_decel
-  # this helps overweighting a_rel when v_lead is close to zero.
-  t_decel = 2.
-  a_rel = min(a_rel, v_lead/t_decel)
-
-  delta = v_rel**2 + 2 * d_rel * a_rel
-  # assign an arbitrary high ttc value if there is no solution to ttc
-  if delta < 0.1:
-    ttc = 5.
-  elif math.sqrt(delta) + v_rel < 0.1:
-    ttc = 5.
-  else:
-    ttc = 2 * d_rel / (math.sqrt(delta) + v_rel)
-  return ttc
-
 
 MAX_SPEED_POSSIBLE = 55.
 
@@ -249,8 +231,9 @@ def compute_speed_with_leads(v_ego, angle_steers, v_pid, l1, l2, CP):
 
   #*** set accel limits as cruise accel/decel limits ***
   a_target = calc_cruise_accel_limits(v_ego)
-  # Always 1 for now.
-  a_pcm = 1
+
+  # start with 1
+  a_pcm = 1.
 
   #*** limit max accel in sharp turns
   a_target, a_pcm = limit_accel_in_turns(v_ego, angle_steers, a_target, a_pcm, CP)
@@ -299,18 +282,14 @@ def compute_speed_with_leads(v_ego, angle_steers, v_pid, l1, l2, CP):
 
 
 class AdaptiveCruise(object):
-  def __init__(self, live20):
-    self.live20 = live20
+  def __init__(self):
     self.last_cal = 0.
     self.l1, self.l2 = None, None
-    self.logMonoTime = 0
     self.dead = True
-  def update(self, cur_time, v_ego, angle_steers, v_pid, CP):
-    l20 = messaging.recv_sock(self.live20)
+  def update(self, cur_time, v_ego, angle_steers, v_pid, CP, l20):
     if l20 is not None:
       self.l1 = l20.live20.leadOne
       self.l2 = l20.live20.leadTwo
-      self.logMonoTime = l20.logMonoTime
 
       # TODO: no longer has anything to do with calibration
       self.last_cal = cur_time